Make `address` field of name server optional
Allow the address to be supplied by the user for cases where it cannot
be computed for the NS entry on apply or when the administrator wish to
use a different value such as "255.255.255.255" for the address. If the
administrator provides an address, it will be used. If not, then the
address will be looked up using the `name` attribute.